changed color

+++ src/main/java/com/el/calculator/models/Calculator.java
@@ -1,26 +1,26 @@
 package com.el.calculator.models;
 
 public class Calculator {
-public int calculator(int firstNumber, int secondNumber, String operation) {
+public double calculator(double firstNumber, double secondNumber, String operation) {
 	if(operation.equals("+")) {
-		int result = firstNumber + secondNumber;
+		double result = firstNumber + secondNumber;
 		return result;
 	}
 	System.out.println(operation);
 	if( operation.equals("-")) {
-		int result = firstNumber - secondNumber;
+		double result = firstNumber - secondNumber;
 		return result;
 	}
 	if( operation.equals("*")) {
-		int result = firstNumber * secondNumber;
+		double result = firstNumber * secondNumber;
 		return result;
 	}
 	if( operation.equals("/")) {
-		int result = firstNumber / secondNumber;
+		double result = firstNumber / secondNumber;
 		return result;
 	}
 	else{
-		int result = (int) Math.pow(firstNumber,secondNumber);
+		double result = (int) Math.pow(firstNumber,secondNumber);
 		return result;
 	}
